requirements.txt
opencv-python
pyserial
mediapipe==0.10.13